requirements.txt
torch>=2.0.0
torchvision>=0.15.0
torchaudio>=2.0.0
pygame>=2.5.0
python-chess>=1.9.0
numpy>=1.24.0 